Modificari filtru, errorController, BarModel, View-uri

+++ nightClub.Domain/Entities/Bar/PhotoBar.cs
@@ -12,8 +12,8 @@ namespace nightClub.Domain.Entities.Bar
         public string Url { get; set; }
         public string Title { get; set; }
         public string Category { get; set; }
-        public int Alcohol { get; set; }
-        public int Price { get; set; }
+        public decimal Alcohol { get; set; }
+        public decimal Price { get; set; }
         public string Description { get; set; }
         public DateTime Date { get; set; }
     }
